requirements.txt
alembic==1.0.0
amqp==2.2.2
aniso8601==3.0.2
anyjson==0.3.3
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
Django==2.0.6
flasgger==0.9.0
Flask==1.0.2
Flask-BabelEx==0.9.3
flask-blueprint==1.2.2
Flask-Cors==3.0.6
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Principal==0.4.0
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
meld3==1.0.2
mistune==0.8.3
mongoengine==0.15.0
msgpack==0.5.6
msgpack-python==0.4.8
paramiko==2.4.1
passlib==1.7.1
pyasn1==0.4.4
pyconvert==0.6.3
pycparser==2.18
PyJWT==1.4.2
pymongo==3.6.1
PyMySQL==0.8.0
PyNaCl==1.2.1
pyrestful==0.4.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2017.3
PyYAML==3.13
qcloudapi-sdk-python==2.0.11
redis==2.10.6
requests==2.18.4
simplejson==3.16.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.1.15
tornado==4.5.2
urllib3==1.22
vine==1.1.4
voluptuous==0.10.5
Werkzeug==0.14.1
WTForms==2.2.1
